Fixes
* Moved local change util functions out of dao
* Moved timestamp function to util
* Use default locale in timestamps
