+++ fhirengine/src/main/java/com/google/fhirengine/sync/ResourceSynchronizer.kt
@@ -16,17 +16,13 @@
 
 package com.google.fhirengine.sync
 
-import android.annotation.SuppressLint
 import com.google.fhirengine.db.Database
 import com.google.fhirengine.db.impl.entities.SyncedResourceEntity
 import com.google.fhirengine.sync.SyncData.Companion.LAST_UPDATED_ASC_VALUE
 import com.google.fhirengine.sync.SyncData.Companion.LAST_UPDATED_KEY
 import com.google.fhirengine.sync.SyncData.Companion.SORT_KEY
+import com.google.fhirengine.toTimeZoneString
 import java.io.IOException
-import java.time.ZoneId
-import java.time.format.DateTimeFormatter
-import java.util.Date
-import java.util.Locale
 import org.hl7.fhir.r4.model.Bundle
 
 /**
@@ -97,14 +93,5 @@ class ResourceSynchronizer(
                 resources
             )
         }
-    }
-
-    @SuppressLint("NewApi") // Use Java8 APIs with desugaring
-    private fun Date.toTimeZoneString(): String {
-        val simpleDateFormat = DateTimeFormatter.ofPattern(
-            "yyyy-MM-dd'T'HH:mm:ss.SSSXXX",
-            Locale.US
-        ).withZone(ZoneId.systemDefault())
-        return simpleDateFormat.format(this.toInstant())
     }
 }
